di thi

+++ app/src/main/java/com/example/sp23_pro1121_cp18103_group4/Adapter/DonHangAdapter.java
@@ -53,7 +53,7 @@ public class DonHangAdapter extends RecyclerView.Adapter<DonHangAdapter.MyViewHo
                 bundle.putString("dhctSoDT",donHang.getSoDt());
                 bundle.putString("dhctDiaChi",donHang.getDiaChi());
                 bundle.putString("dhctNgay",donHang.getNgayThanhToan());
-                bundle.putInt("dhctTongTien",donHang.getTongTien());
+                bundle.putFloat("dhctTongTien",donHang.getTongTien());
                 bundle.putString("dhctTrangThai",donHang.getTrangThai());
                 bundle.putInt("dhctMaDatHang",donHang.getMaDatHang());
                 Fragment fragment = new DonHangChiTietFragment();
